fix : 버그 수정

+++ Assets/Script/UI/Sub/UI_HPBar.cs
@@ -53,7 +53,7 @@ public class UI_HPBar : UI_Base
             _enemyBar.fillAmount = amount;
     }
 
-    public void RefreshFallGauge(float current)
+    public void RefreshFallGauge(int current)
     {
         for (int i = 0; i < current; i++)
             Util.FindChild(_fallGauge[i].gameObject, "Fill").SetActive(true);
